Replaced many types with 64bit-ready equivalents

+++ Source/MusicServer/Artist.m
@@ -77,7 +77,7 @@ NSString *gUnknownArtistName = @"gUnknownArtistName";
 	self = [super init];
 	mName = [[decoder decodeObject] retain];
 	
-	unsigned length;
+	NSUInteger length;
 	memcpy(&_identifier, [decoder decodeBytesWithReturnedLength:&length], sizeof(_identifier));
 	
 	return self;
